+++ BusinessLogic/Contracts/IVehicleService.cs
@@ -8,6 +8,6 @@ namespace BusinessLogic.Contracts
     public interface IVehicleService : IFormService<IVehicleRepository, Vehicle>
     {
         Vehicle AddOrUpdate(VehicleViewModel model);
-        new IEnumerable<Vehicle> GetByUserId(int userId);
+        new IEnumerable<Vehicle> GetByUserId(long userId);
     }
 }
