+++ Models/Contract/IFormEntity.cs
@@ -3,7 +3,7 @@
     public interface IFormEntity        
     {
         long Id { get; set; }
-        int UserId { get; set; }
+        long UserId { get; set; }
         IViewModel ConvertToModel();
         void Update(IFormEntity entity);
     }
